+++ src/main/java/elasticsearch/sample/update/UpdateMapDataRenewal.java
@@ -1,4 +1,4 @@
-package elasticsearch.sample;
+package elasticsearch.sample.update;
 
 import java.io.IOException;
 import java.util.HashMap;
